Moving selector logic to its own package [skip ci]
