+++ java/server/src/org/openqa/selenium/grid/distributor/selector/BUILD.bazel
@@ -1,0 +1,23 @@
+load("@rules_jvm_external//:defs.bzl", "artifact")
+
+java_library(
+    name = "selector",
+    srcs = glob(["*.java"]),
+    visibility = [
+        "//java/server/src/org/openqa/selenium:__subpackages__",
+        "//java/server/test/org/openqa/selenium:__subpackages__",
+    ],
+    deps = [
+      "//java/server/src/org/openqa/selenium/grid/data",
+      "//java/server/src/org/openqa/selenium/grid/node",
+      "//java/server/src/org/openqa/selenium/grid/node/remote",
+      "//java/server/src/org/openqa/selenium/grid/component",
+      "//java/server/src/org/openqa/selenium/events",
+      "//java/client/src/org/openqa/selenium:core",
+      "//java/client/src/org/openqa/selenium/remote",
+      "//java/server/src/org/openqa/selenium/concurrent",
+      "//java/client/src/org/openqa/selenium/json",
+      "//java/server/src/org/openqa/selenium/grid/distributor/model",
+      artifact("com.google.guava:guava"),
+    ],
+)
