+++ java/server/test/org/openqa/selenium/grid/distributor/selector/BUILD.bazel
@@ -16,6 +16,7 @@ java_test_suite(
         "//java/server/src/org/openqa/selenium/grid/data",
         "//java/client/src/org/openqa/selenium/remote/http",
         "//java/server/src/org/openqa/selenium/grid/distributor/model",
+        "//java/server/src/org/openqa/selenium/grid/distributor/selector",
         artifact("junit:junit"),
         artifact("org.assertj:assertj-core"),
     ]
